requirements.txt

transformers
torch
sentence-transformers
faiss-cpu
rank-bm25
pypdf
numpy
streamlit
accelerate
huggingface_hub
login
ollama
pdfplumber
scikit-learn
pymupdf
langchain-text-splitters
gradio
watchdog
python-dotenv